Refactoring step 5 - move things around
