+++ haystack_experimental/components/tools/openapi/schema_conversion.py
@@ -4,21 +4,18 @@
 
 import logging
 from typing import Any, Callable, Dict, List, Optional
+
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-
-
 import jsonref
-
-VALID_HTTP_METHODS = ["get", "put", "post", "delete", "options", "head", "patch", "trace"]
 
 MIN_REQUIRED_OPENAPI_SPEC_VERSION = 3
 
 logger = logging.getLogger(__name__)
 
 
-def openai_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]: # noqa: F821
+def openai_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]:  # noqa: F821
     """
     Converts OpenAPI specification to a list of function suitable for OpenAI LLM function calling.
 
@@ -26,11 +23,13 @@ def openai_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]: # 
     :return: A list of dictionaries, each representing a function definition.
     """
     resolved_schema = jsonref.replace_refs(schema.spec_dict)
-    fn_definitions = _openapi_to_functions(resolved_schema, "parameters", _parse_endpoint_spec_openai)
+    fn_definitions = _openapi_to_functions(
+        resolved_schema, "parameters", _parse_endpoint_spec_openai
+    )
     return [{"type": "function", "function": fn} for fn in fn_definitions]
 
 
-def anthropic_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]: # noqa: F821
+def anthropic_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]:  # noqa: F821
     """
     Converts an OpenAPI specification to a list of function definitions for Anthropic LLM function calling.
 
@@ -38,10 +37,12 @@ def anthropic_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]:
     :return: A list of dictionaries, each representing a function definition.
     """
     resolved_schema = jsonref.replace_refs(schema.spec_dict)
-    return _openapi_to_functions(resolved_schema, "input_schema", _parse_endpoint_spec_openai)
+    return _openapi_to_functions(
+        resolved_schema, "input_schema", _parse_endpoint_spec_openai
+    )
 
 
-def cohere_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]: # noqa: F821
+def cohere_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]:  # noqa: F821
     """
     Converts an OpenAPI specification to a list of function definitions for Cohere LLM function calling.
 
@@ -49,7 +50,9 @@ def cohere_converter(schema: "OpenAPISpecification") -> List[Dict[str, Any]]: # 
     :return: A list of dictionaries, each representing a function definition.
     """
     resolved_schema = jsonref.replace_refs(schema.spec_dict)
-    return _openapi_to_functions(resolved_schema, "not important for cohere", _parse_endpoint_spec_cohere)
+    return _openapi_to_functions(
+        resolved_schema, "not important for cohere", _parse_endpoint_spec_cohere
+    )
 
 
 def _openapi_to_functions(
@@ -65,7 +68,9 @@ def _openapi_to_functions(
     # We check the version and require minimal fields to be present, so we can extract functions
     spec_version = service_openapi_spec.get("openapi")
     if not spec_version:
-        raise ValueError(f"Invalid OpenAPI spec provided. Could not extract version from {service_openapi_spec}")
+        raise ValueError(
+            f"Invalid OpenAPI spec provided. Could not extract version from {service_openapi_spec}"
+        )
     service_openapi_spec_version = int(spec_version.split(".")[0])
     # Compare the versions
     if service_openapi_spec_version < MIN_REQUIRED_OPENAPI_SPEC_VERSION:
@@ -82,19 +87,26 @@ def _openapi_to_functions(
     return functions
 
 
-def _parse_endpoint_spec_openai(resolved_spec: Dict[str, Any], parameters_name: str) -> Dict[str, Any]:
+def _parse_endpoint_spec_openai(
+    resolved_spec: Dict[str, Any], parameters_name: str
+) -> Dict[str, Any]:
     """
     Parses an OpenAPI endpoint specification for OpenAI.
     """
     if not isinstance(resolved_spec, dict):
-        logger.warning("Invalid OpenAPI spec format provided. Could not extract function.")
+        logger.warning(
+            "Invalid OpenAPI spec format provided. Could not extract function."
+        )
         return {}
     function_name = resolved_spec.get("operationId")
     description = resolved_spec.get("description") or resolved_spec.get("summary", "")
     schema: Dict[str, Any] = {"type": "object", "properties": {}}
     # requestBody section
     req_body_schema = (
-        resolved_spec.get("requestBody", {}).get("content", {}).get("application/json", {}).get("schema", {})
+        resolved_spec.get("requestBody", {})
+        .get("content", {})
+        .get("application/json", {})
+        .get("schema", {})
     )
     if "properties" in req_body_schema:
         for prop_name, prop_schema in req_body_schema["properties"].items():
@@ -108,14 +120,23 @@ def _parse_endpoint_spec_openai(resolved_spec: Dict[str, Any], parameters_name: 
             schema_dict = _parse_property_attributes(param["schema"])
             # these attributes are not in param[schema] level but on param level
             useful_attributes = ["description", "pattern", "enum"]
-            schema_dict.update({key: param[key] for key in useful_attributes if param.get(key)})
+            schema_dict.update(
+                {key: param[key] for key in useful_attributes if param.get(key)}
+            )
             schema["properties"][param["name"]] = schema_dict
             if param.get("required", False):
                 schema.setdefault("required", []).append(param["name"])
 
     if function_name and description and schema["properties"]:
-        return {"name": function_name, "description": description, parameters_name: schema}
-    logger.warning("Invalid OpenAPI spec format provided. Could not extract function from %s", resolved_spec)
+        return {
+            "name": function_name,
+            "description": description,
+            parameters_name: schema,
+        }
+    logger.warning(
+        "Invalid OpenAPI spec format provided. Could not extract function from %s",
+        resolved_spec,
+    )
     return {}
 
 
@@ -134,7 +155,8 @@ def _parse_property_attributes(
     if schema_type == "object":
         properties = property_schema.get("properties", {})
         parsed_properties = {
-            prop_name: _parse_property_attributes(prop, include_attributes) for prop_name, prop in properties.items()
+            prop_name: _parse_property_attributes(prop, include_attributes)
+            for prop_name, prop in properties.items()
         }
         parsed_schema["properties"] = parsed_properties
         if "required" in property_schema:
@@ -145,7 +167,9 @@ def _parse_property_attributes(
     return parsed_schema
 
 
-def _parse_endpoint_spec_cohere(operation: Dict[str, Any], ignored_param: str) -> Dict[str, Any]:
+def _parse_endpoint_spec_cohere(
+    operation: Dict[str, Any], ignored_param: str
+) -> Dict[str, Any]:
     """
     Parses an endpoint specification for Cohere.
     """
@@ -170,19 +194,27 @@ def _parse_parameters(operation: Dict[str, Any]) -> Dict[str, Any]:
     for param in operation.get("parameters", []):
         if "schema" in param:
             parameters[param["name"]] = _parse_schema(
-                param["schema"], param.get("required", False), param.get("description", "")
+                param["schema"],
+                param.get("required", False),
+                param.get("description", ""),
             )
     if "requestBody" in operation:
-        content = operation["requestBody"].get("content", {}).get("application/json", {})
+        content = (
+            operation["requestBody"].get("content", {}).get("application/json", {})
+        )
         if "schema" in content:
             schema_properties = content["schema"].get("properties", {})
             required_properties = content["schema"].get("required", [])
             for name, schema in schema_properties.items():
-                parameters[name] = _parse_schema(schema, name in required_properties, schema.get("description", ""))
+                parameters[name] = _parse_schema(
+                    schema, name in required_properties, schema.get("description", "")
+                )
     return parameters
 
 
-def _parse_schema(schema: Dict[str, Any], required: bool, description: str) -> Dict[str, Any]:  # noqa: FBT001
+def _parse_schema(
+    schema: Dict[str, Any], required: bool, description: str
+) -> Dict[str, Any]:  # noqa: FBT001
     """
     Parses a schema part of an operation specification.
     """
