+++ test/components/tools/openapi/__init__.py
@@ -1,7 +1,3 @@
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-
-from .openai.function_caller import OpenAIFunctionCaller
-
-_all_ = ["OpenAIFunctionCaller"]
