+++ test/components/tools/openapi/test_openapi_client.py
@@ -7,8 +7,8 @@ from fastapi import FastAPI
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
 
-from haystack_experimental.util.openapi import OpenAPIServiceClient, ClientConfiguration
-from test.util.conftest import FastAPITestClient
+from haystack_experimental.components.tools.openapi.openapi import OpenAPIServiceClient, ClientConfiguration
+from test.components.tools.openapi.conftest import FastAPITestClient
 
 """
 Tests OpenAPIServiceClient with three FastAPI apps for different parameter types:
